+++ blueprint/implementations/worker/bullmq/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forklaunch/implementation-worker-bullmq
 
+## 0.8.21
+
+### Patch Changes
+
+- internal bump
+- Updated dependencies
+  - @forklaunch/interfaces-worker@0.7.21
+
 ## 0.8.20
 
 ### Patch Changes