+++ blueprint/interfaces/billing/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/interfaces-billing
 
+## 0.8.21
+
+### Patch Changes
+
+- internal bump
+
 ## 0.8.20
 
 ### Patch Changes