+++ framework/universal-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forklaunch/universal-sdk
 
+## 0.7.36
+
+### Patch Changes
+
+- internal package bump
+- Updated dependencies
+  - @forklaunch/common@0.6.36
+
 ## 0.7.35
 
 ### Patch Changes